+++ syndicate/core/generators/deployment_resources/step_function_generator.py
@@ -18,4 +18,6 @@ class StepFunctionGenerator(BaseDeploymentResourceGenerator):
         "event_sources": list,
         "dependencies": list,
         "iam_role": None,
+        "publish_version": bool,
+        "alias": None
     }