+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.12.0] - 2024-05-30
+- Raised version of libraries `boto3` and `botocore`
+- Added support of new features to the `step_functions` resource:
+  * update
+  * publish version
+  * state machine aliases
+
 # [1.11.6] - 2024-05-24
 - Added support of custom authorizer names in Open API specification security schemes
 - Fixed quietness of errors while deploying/updating API Gateway via OpenAPI specification